--- opentelemetry-aws/src/trace/xray/mod.rs
@@ -1,1 +1,0 @@
-pub mod propagator;
